+++ config/config.exs
@@ -407,6 +407,8 @@ config :pleroma, :mrf_vocabulary,
   accept: [],
   reject: []
 
+config :pleroma, :mrf_inline_quote, prefix: "RE"
+
 # threshold of 7 days
 config :pleroma, :mrf_object_age,
   threshold: 604_800,