+++ lib/pleroma/web/api_spec/operations/status_operation.ex
@@ -496,6 +496,11 @@ defmodule Pleroma.Web.ApiSpec.StatusOperation do
           type: :string,
           description:
             "Will reply to a given conversation, addressing only the people who are part of the recipient set of that conversation. Sets the visibility to `direct`."
+        },
+        quote_id: %Schema{
+          nullable: true,
+          type: :string,
+          description: "Will quote a given status."
         }
       },
       example: %{