+++ lib/pleroma/web/common_api/activity_draft.ex
@@ -22,6 +22,8 @@ defmodule Pleroma.Web.CommonAPI.ActivityDraft do
             attachments: [],
             in_reply_to: nil,
             in_reply_to_conversation: nil,
+            quote_id: nil,
+            quote: nil,
             visibility: nil,
             expires_at: nil,
             extra: nil,
@@ -54,6 +56,7 @@ defmodule Pleroma.Web.CommonAPI.ActivityDraft do
     |> with_valid(&in_reply_to/1)
     |> with_valid(&in_reply_to_conversation/1)
     |> with_valid(&visibility/1)
+    |> with_valid(&quote_id/1)
     |> content()
     |> with_valid(&to_and_cc/1)
     |> with_valid(&context/1)
@@ -107,6 +110,28 @@ defmodule Pleroma.Web.CommonAPI.ActivityDraft do
     in_reply_to_conversation = Participation.get(draft.params[:in_reply_to_conversation_id])
     %__MODULE__{draft | in_reply_to_conversation: in_reply_to_conversation}
   end
+
+  defp quote_id(%{params: %{quote_id: ""}} = draft), do: draft
+
+  defp quote_id(%{params: %{quote_id: id}} = draft) when is_binary(id) do
+    with {:activity, %Activity{} = quote} <- {:activity, Activity.get_by_id(id)},
+         visibility <- CommonAPI.get_quoted_visibility(quote),
+         {:visibility, true} <- {:visibility, visibility in ["public", "unlisted"]} do
+      %__MODULE__{draft | quote: Activity.get_by_id(id)}
+    else
+      {:activity, _} ->
+        add_error(draft, dgettext("errors", "You can't quote a status that doesn't exist"))
+
+      {:visibility, false} ->
+        add_error(draft, dgettext("errors", "You can only quote public or unlisted statuses"))
+    end
+  end
+
+  defp quote_id(%{params: %{quote_id: %Activity{} = quote}} = draft) do
+    %__MODULE__{draft | quote: quote}
+  end
+
+  defp quote_id(draft), do: draft
 
   defp visibility(%{params: params} = draft) do
     case CommonAPI.get_visibility(params, draft.in_reply_to, draft.in_reply_to_conversation) do